Add comments about the relation to terratest
